+++ Chat/ChatRoomViewController.swift
@@ -1,0 +1,21 @@
+//
+//  ChatRoomViewController.swift
+//  Chat
+//
+//  Created by Hanawa Takuro on 2016/09/11.
+//  Copyright © 2016年 Hanawa Takuro. All rights reserved.
+//
+
+import UIKit
+
+class ChatRoomViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+    }
+}
