+++ Podfile
@@ -1,5 +1,5 @@
 # Uncomment this line to define a global platform for your project
-# platform :ios, '9.0'
+platform :ios, '9.0'
 
 target 'Chat' do
   # Comment this line if you're not using Swift and don't want to use dynamic frameworks
@@ -10,3 +10,10 @@ target 'Chat' do
   pod 'AWSCognito'
 
 end
+
+plugin 'cocoapods-keys', {
+  :project => "Chat",
+  :keys => [
+    "CognitoIdentityPoolId"
+  ]
+}
